feat: ProjectList 컴포넌트 및 스타일 파일별 분리
