+++ src/views/ProjectPage/styles/desktop-filter.module.scss
@@ -1,0 +1,52 @@
+.project-filter {
+  position: fixed;
+  top: 100px;
+  width: 250px;
+  height: 700px;
+
+  background: #181818;
+  border-radius: 10px 10px 0px 0px;
+  padding: 32px 32px 32px 32px;
+
+  .title {
+    display: flex;
+    align-items: center;
+    margin-bottom: 30px;
+
+    h4 {
+      font-size: 20px;
+      color: #ffffff;
+      padding-top: 4px;
+      padding-left: 10px;
+    }
+  }
+
+  .content {
+    width: 100%;
+    height: 60%;
+    z-index: 2;
+
+    display: flex;
+    flex-direction: column;
+    gap: 25px;
+
+    span {
+      display: flex;
+      justify-content: flex-start;
+      align-items: center;
+
+      color: #ffffff;
+      font-size: 17px;
+
+      width: 140px;
+      height: 80px;
+      border-radius: 10px;
+      padding-left: 18px;
+      cursor: pointer;
+
+      &.isClicked {
+        background: rgba(255, 255, 255, 0.05);
+      }
+    }
+  }
+}
