+++ src/views/ProjectPage/styles/mobile-filter.module.scss
@@ -1,0 +1,100 @@
+.utility {
+  position: fixed;
+  top: 80%;
+  right: 20px;
+  z-index: 100;
+
+  display: flex;
+  flex-direction: column;
+  gap: 10px 0;
+
+  .up {
+    width: 56px;
+    height: 56px;
+    border-radius: 50%;
+
+    background: #242424;
+    box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.25);
+  }
+
+  .filter {
+    width: 56px;
+    height: 56px;
+    border-radius: 50%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+
+    background: #ffffff;
+    box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.25);
+  }
+}
+
+// Filter Modal
+.modal {
+  position: fixed;
+  top: 0;
+  z-index: 99;
+  width: 100vw;
+  height: 100vh;
+
+  border-radius: 10px 10px 0px 0px;
+
+  .overlay {
+    width: 100%;
+    height: 50%;
+    cursor: pointer;
+
+    opacity: 0.8;
+    background-color: rgba(0, 0, 0, 0.5);
+  }
+
+  .content {
+    width: 100%;
+    height: 50%;
+    background: #181818;
+    border-radius: 10px 10px 0px 0px;
+    padding: 32px 32px 0 32px;
+
+    .title {
+      display: flex;
+      align-items: center;
+      margin-bottom: 30px;
+
+      h4 {
+        font-size: 18px;
+        color: #ffffff;
+        padding-top: 4px;
+        padding-left: 10px;
+      }
+
+      svg.isRotated {
+        transform: rotate(90deg);
+      }
+    }
+
+    .categories {
+      display: grid;
+      grid-template-columns: repeat(2, 1fr);
+      gap: 20px;
+
+      span {
+        display: flex;
+        justify-content: flex-start;
+        align-items: center;
+
+        color: #ffffff;
+        font-size: 16px;
+
+        width: 140px;
+        height: 40px;
+        border-radius: 10px;
+        padding-left: 18px;
+
+        &.isClicked {
+          background: rgba(255, 255, 255, 0.05);
+        }
+      }
+    }
+  }
+}
