+++ src/views/ProjectPage/styles/project-enroll.module.scss
@@ -1,0 +1,28 @@
+.enroll {
+  height: 402px;
+
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  font-size: 20px;
+
+  @media (max-width: 1279px) {
+    height: 250px;
+    font-size: 15px;
+  }
+
+  p {
+    padding-bottom: 10px;
+    color: rgba(255, 255, 255, 0.5);
+  }
+
+  span {
+    color: rgba(255, 255, 255, 0.5);
+    font-weight: 700;
+  }
+
+  div {
+    cursor: pointer;
+  }
+}
