+++ src/views/ProjectPage/styles/project-list.module.scss
@@ -1,0 +1,57 @@
+.total-container {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  justify-content: center;
+
+  align-items: flex-end;
+  @media (max-width: 1279px) {
+    align-items: center;
+  }
+}
+
+.list-container {
+  width: 90%;
+  height: 100%;
+
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  margin-top: 130px;
+
+  @media (max-width: 1279px) {
+    margin-top: 90px;
+  }
+
+  .total-count {
+    display: flex;
+    align-items: center;
+    margin: 5px 0 30px 0;
+
+    p {
+      color: #8e8e8e;
+      font-size: 24px;
+
+      @media screen and (max-width: 600px) {
+        font-size: 18px;
+      }
+    }
+  }
+
+  .list-data-container {
+    width: 80%;
+
+    display: grid;
+    grid-template-columns: 1fr 1fr 1fr;
+    gap: 20px;
+
+    @media screen and (max-width: 600px) {
+      display: flex;
+      flex-direction: column;
+      align-items: center;
+      justify-content: center;
+      gap: 0;
+    }
+  }
+}
