+++ src/views/ProjectPage/styles/project.module.scss
@@ -1,0 +1,84 @@
+.item {
+  width: 328px;
+  height: 402px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+
+  border: 1px solid #353535;
+  border-radius: 10px;
+  margin-bottom: 20px;
+  cursor: pointer;
+
+  .image-wrapper {
+    width: 316px;
+    height: 180px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+  }
+
+  .content {
+    width: 90%;
+    height: 220px;
+
+    display: flex;
+    flex-direction: column;
+    justify-content: space-around;
+
+    .types {
+      display: flex;
+      justify-content: flex-start;
+
+      & > div {
+        display: flex;
+        justify-content: center;
+        align-items: center;
+
+        width: 50px;
+        height: 32px;
+        padding: 10px 10px;
+
+        font-size: 12px;
+
+        border: 1px solid rgba(255, 255, 255, 0.3);
+        border-radius: 10px;
+        color: #cccccc;
+
+        margin-right: 8px;
+      }
+    }
+
+    .text {
+      height: 50px;
+
+      & > h5 {
+        font-size: 20px;
+      }
+
+      & > p {
+        margin-top: 10px;
+        color: #8e8e8e;
+      }
+    }
+
+    .links {
+      & > div {
+        display: flex;
+      }
+
+      .icon-wrapper {
+        display: flex;
+        justify-content: center;
+        align-items: center;
+
+        width: 50px;
+        height: 50px;
+
+        background: #242424;
+        border-radius: 5px;
+        margin-right: 10px;
+      }
+    }
+  }
+}
